feat: Agrega  3 pruebas adicionales para la tarea de Pytest semana 3
